+++ src/main/java/com/example/aucaregistration/repository/TicketRepository.java
@@ -3,4 +3,4 @@ package com.example.aucaregistration.repository;
 import com.example.aucaregistration.domain.Ticket;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface TicketRepository extends JpaRepository<Ticket, Long> { }
+public interface TicketRepository extends JpaRepository<Ticket, Integer> { }
